LPS-75906 Check for existence before deletion
User notification events are automatically removed as soon as it is
detected that the thing it is notifying about (e.g. Document, Web
content, etc) has been deleted. See for example
BaseModelUserNotificationHandler#getBody (around line 85).

To avoid changing the behaviour of base classes, I'm handling this
special case in the portlet.

+++ modules/apps/collaboration/notifications/notifications-web/src/main/java/com/liferay/notifications/web/internal/portlet/NotificationsPortlet.java
@@ -17,6 +17,7 @@ package com.liferay.notifications.web.internal.portlet;
 import com.liferay.notifications.web.internal.constants.NotificationsPortletKeys;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.Release;
+import com.liferay.portal.kernel.model.UserNotificationDelivery;
 import com.liferay.portal.kernel.model.UserNotificationDeliveryConstants;
 import com.liferay.portal.kernel.model.UserNotificationEvent;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
@@ -101,8 +102,14 @@ public class NotificationsPortlet extends MVCPortlet {
 			actionRequest, "userNotificationEventId");
 
 		try {
-			_userNotificationEventLocalService.deleteUserNotificationEvent(
-				userNotificationEventId);
+			UserNotificationEvent userNotificationEvent =
+				_userNotificationEventLocalService.
+					fetchUserNotificationEvent(userNotificationEventId);
+
+			if (userNotificationEvent != null) {
+				_userNotificationEventLocalService.deleteUserNotificationEvent(
+					userNotificationEvent);
+			}
 		}
 		catch (Exception e) {
 			throw new PortletException(e);
